src/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
alembic
bcrypt
PyJWT
click
passlib
Pillow
